🗃️ ftree: move InsigniasTable files into InsigniasPage
