+++ apps/studio/src/ui/shop/widgets/pages/InsigniasPage/InsigniasTable/DeleteInsigniaDialog/DeleteInsigniaDialogView.tsx
