+++ apps/studio/src/ui/shop/widgets/pages/InsigniasPage/InsigniasTable/InsigniaForm/index.tsx
@@ -1,0 +1,1 @@
+export { InsigniaFormView as InsigniaForm } from './InsigniaFormView'
