+++ apps/studio/src/ui/shop/widgets/pages/InsigniasPage/InsigniasTable/useInsigniasTable.ts
