+++ apps/studio/src/ui/shop/widgets/pages/InsigniasPage/index.tsx
@@ -1,0 +1,1 @@
+export { InsigniasPageView as InsigniasPage } from './InsigniasPageView'
