--- apps/studio/src/ui/shop/widgets/pages/Shop/InsigniasTable/InsigniaForm/index.tsx
@@ -1,1 +1,0 @@
-export { InsigniaFormView } from './InsigniaFormView'
